Fix compiler warning fail in build-fc10 (hopefully)

+++ src/ee/storage/table.h
@@ -361,7 +361,8 @@ protected:
     virtual void onSetColumns() {};
 
     double loadFactor() {
-        return activeTupleCount() / static_cast<double>(allocatedTupleCount());
+        return static_cast<double>(activeTupleCount()) /
+            static_cast<double>(allocatedTupleCount());
     }
 
     // TUPLES AND MEMORY USAGE
